Fix warnings
Former-commit-id: d3876b1121d5b2b3f91baf1e1f40a5395b6f378d

+++ CammentSDK/Classes/Modules/CammentView/Interactor/CMCammentViewInteractor.m
@@ -191,7 +191,7 @@ NSString *const CMCammentViewInteractorErrorDomain = @"tv.camment.CMCammentViewI
         return;
     }
     
-    DDLogVerbose(@"Start uploading attempt %d", nextAttemptNumber);
+    DDLogVerbose(@"Start uploading attempt %ld", (long)nextAttemptNumber);
     [self uploadCamment:camment];
 }
